Authentication_ZAP_GT_CRIVO/framework/requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
exceptiongroup==1.2.1
h11==0.14.0
outcome==1.3.0.post0
pydantic==2.8.2
pydantic_core==2.20.1
PySocks==1.7.1
PyYAML==6.0.1
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
selenium==4.22.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
websocket-client==1.8.0
wsproto==1.2.0
zaproxy==0.3.2
